requirements.txt
pyspark==3.3.0
delta-spark==2.1.0
pytest==7.3.1